+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
 
+## 0.0.2 (November 7, 2018)
+
++ `F8` run selection and delete
++ error log
++ workspace log
++ log when run scripts
++ highlight setting for `%%` `% [...]`
+
++ fix: clear before run
++ fix: `F9` support `...`breaking line
++ fix: support `import`
+
 ## 0.0.1 (November 1, 2018)
 
 Finish first publish. 
 
-+ support `F5` `F9` `Ctrl+Enter`
-
-+ log when run scripts
++ support `F5` `F9` `Ctrl+Enter`  
+  `F9` run current line when empty selection insteadof do nothing